+++ src/IconFactory/index.ts
@@ -1,4 +1,0 @@
-import type { IconName } from './IconFactory'
-
-export { default } from '././IconFactory'
-export type { IconName }